+++ lua/haskell-tools/lsp/definition.lua
@@ -30,7 +30,6 @@ M.setup = function(config)
       end
     end
   end
-
 end
 
 return M