+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Command for loading files into the repl.
+- Ability to paste multiple valid Haskell lines into the repl.
 
 ## [1.2.0] - 2022-11-09
 ### Added