+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Command for loading files into the repl.
 - Ability to paste multiple valid Haskell lines into the repl.
+- `repl.paste_info` and `repl.cword_info` functions.
 
 ## [1.2.0] - 2022-11-09
 ### Added