+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Command for loading files into the repl.
 - Ability to paste multiple valid Haskell lines into the repl.
 - `repl.paste_info` and `repl.cword_info` functions.
+- Telescope live_grep and find_files commands for current package
+- Register Telescope extension
 ### Fixed
 - Fix broken `<C-b>` keymap to open Hoogle entry in the browser