+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Configuration for falling back to hoogle search if 'vim.lsp.definition' fails
+### Fixed
+- Hover actions improvements:
+  - Always show 'Go to definition' if location is found
+  - Fall back to hoogle search if 'Go to definition' fails
 
 ## [1.3.0] - 2022-11-14
 ### Added