+++ CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Configuration for falling back to hoogle search if 'vim.lsp.definition' fails
+- Configuration for falling back to hoogle search if 'vim.lsp.definition' fails.
 ### Fixed
 - Hover actions improvements:
-  - Always show 'Go to definition' if location is found
-  - Offer Hoogle search for package <> name if location is not found
+  - Always show 'Go to definition' if location is found.
+  - Offer Hoogle search for package <> name if location is not found.
+- Hoogle search: Replace multiple whitespace with single space.
 
 ## [1.3.0] - 2022-11-14
 ### Added