+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Automatically generate project & package tags if [`fast-tags`](https://github.com/elaforge/fast-tags) is installed.
 - Configuration for falling back to hoogle search if 'vim.lsp.definition' fails.
-- flake setup
+- Nix flake setup
 ### Fixed
 - Hover actions improvements:
   - Always show 'Go to definition' if location is found.