+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.0] - 2022-11-18
 ### Added
 - Automatically generate project & package tags if [`fast-tags`](https://github.com/elaforge/fast-tags) is installed.
 - Configuration for falling back to hoogle search if 'vim.lsp.definition' fails.
-- Nix flake setup
+- Nix flake setup.
 ### Fixed
 - Hover actions improvements:
   - Always show 'Go to definition' if location is found.